Switched to Zeitwerk as autoloader.
